[AakashMallik] UI refactor complete
