+++ client/src/pages/settings/components/text-editor-settings.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { textEditorSettings as settings } from '../data/text-editor-settings'
+import SettingsGroup from './settings-group'
+
+export default function TextEditorSettings() {
+  const subGroupDict = settings.subGroupDict
+
+  return (
+    <>
+      {Object.keys(subGroupDict).map((subGroupKeys, idx) => (
+        <SettingsGroup key={idx} subGroup={subGroupDict[subGroupKeys]} />
+      ))}
+    </>
+  )
+}
