+++ docker-compose.yml
@@ -33,25 +33,25 @@ services:
       - postgres
 
   # backend prod
-  server_prod:
-    domainname: backend
-    stdin_open: true
-    build:
-      context: ./server
-      dockerfile: Dockerfile.prod
-    ports:
-      - '3001:3001'
-    volumes:
-      - './server:/srv/server'
-    environment:
-      - PORT=3001
-      - POSTGRES_DB=${POSTGRES_DB}
-      - DB_HOST=${DB_HOST}
-      - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
-      - POSTGRES_USER=${POSTGRES_USER}
-      - POSTGRES_PORT=${POSTGRES_PORT}
-    depends_on:
-      - postgres
+  # server_prod:
+  #   domainname: backend
+  #   stdin_open: true
+  #   build:
+  #     context: ./server
+  #     dockerfile: Dockerfile.prod
+  #   ports:
+  #     - '3001:3001'
+  #   volumes:
+  #     - './server:/srv/server'
+  #   environment:
+  #     - PORT=3001
+  #     - POSTGRES_DB=${POSTGRES_DB}
+  #     - DB_HOST=${DB_HOST}
+  #     - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
+  #     - POSTGRES_USER=${POSTGRES_USER}
+  #     - POSTGRES_PORT=${POSTGRES_PORT}
+  #   depends_on:
+  #     - postgres
 
   # frontend
   client:
